pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = [
    {name = "Bruno-Alvez",email = "bruno.bsantos75@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"

